requirements.txt
lark-parser==0.12.0
z3-solver==4.8.10.0
pydot==1.4.2
pathlib==1.0.1
pytest==6.2.5
flake8==3.9.2
pycco==0.6.0
